% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedisBackend.R
\name{RedisBackend}
\alias{RedisBackend}
\title{Creating the Redis backend}
\usage{
RedisBackend(
  RedisParam = NULL,
  jobname = "",
  host = "127.0.0.1",
  port = 6379L,
  password = NULL,
  timeout = 2592000L,
  type = c("manager", "worker")
)
}
\arguments{
\item{RedisParam}{RedisParam, if this argument is not NULL, all the
other arguments will be ignored except `type`.}

\item{jobname}{character(1) The job name used by the manager and
workers to connect.}

\item{host}{character(1) The host of the Redis server.}

\item{port}{integer(1) The port of the Redis server.}

\item{password}{character(1) The password of the redis server.}

\item{timeout}{integer(1) The waiting time in `BLPOP`.}

\item{type}{character(1) The type of the backend (manager or worker?).}
}
\description{
Creating the Redis backend
}
\keyword{internal}
