% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedisBackend.R
\name{RedisBackend}
\alias{RedisBackend}
\title{Creating the Redis backend}
\usage{
RedisBackend(
  jobname,
  host = "127.0.0.1",
  port = 6379L,
  password = NULL,
  timeout = 2592000L,
  type = c("manager", "worker"),
  RedisParam = NULL
)
}
\arguments{
\item{jobname}{The job name used by the manager and workers to connect.}

\item{host}{The host of the Redis server.}

\item{port}{The port of the Redis server.}

\item{password}{The password of the redis server.}

\item{timeout}{The waiting time in `BLPOP`}

\item{type}{The type of the Backend(manager or worker?).}

\item{RedisParam}{RedisParam, if this argument is not NULL, all the other
arguments will be ignored except `type`.}
}
\description{
Creating the Redis backend
}
