% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedisParam-class.R
\name{bpstopall}
\alias{bpstopall}
\title{Deprecated functions in the RedisParam package}
\usage{
bpstopall(x)
}
\value{
See \code{?rpstopall} for return value.
}
\description{
\code{bpstopall()} is provided for compatibility with
previous versions of RedisParam, and will be defunct after the
next release. Use \code{rpstopall()} instead.
}
\examples{
if (FALSE) {
    ## bpstopall()
    ## deprecated -- use rpstopall() instead
}

}
