% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedisParam-class.R
\docType{methods}
\name{RedisParam}
\alias{RedisParam}
\alias{.send,redis_worker-method}
\alias{.recv,redis_worker-method}
\alias{.close,redis_worker-method}
\alias{.send_to,redis_manager-method}
\alias{.recv_any,redis_manager-method}
\alias{bpisup,RedisParam-method}
\alias{bpbackend,RedisParam-method}
\alias{bpstart,RedisParam-method}
\alias{bpstop,RedisParam-method}
\alias{bpstopall}
\alias{bpstopall,RedisParam-method}
\title{Enable redis-based parallel evaluation in BiocParallel}
\usage{
RedisParam(workers = snowWorkers(), tasks = 0L, jobname = ipcid(),
  log = FALSE, logdir = NA, threshold = "INFO",
  resultdir = NA_character_, stop.on.error = TRUE,
  timeout = 2592000L, exportglobals = TRUE, progressbar = FALSE,
  RNGseed = NULL, manager.hostname = "127.0.0.1",
  manager.port = 6379L, is.worker = NA)

\S4method{bpisup}{RedisParam}(x)

\S4method{bpbackend}{RedisParam}(x)

\S4method{bpstart}{RedisParam}(x, ...)

\S4method{bpstop}{RedisParam}(x)

bpstopall(x)

\S4method{bpstopall}{RedisParam}(x)
}
\arguments{
\item{workers}{See BiocParalleParam-class.}

\item{tasks}{See BiocParalleParam-class.}

\item{jobname}{character(1) name (unique) used to associate manager
& workers on a queue.}

\item{log}{See BiocParalleParam-class.}

\item{logdir}{See BiocParalleParam-class.}

\item{threshold}{See BiocParalleParam-class.}

\item{resultdir}{See BiocParalleParam-class.}

\item{stop.on.error}{See BiocParalleParam-class.}

\item{timeout}{See BiocParalleParam-class.}

\item{exportglobals}{See BiocParalleParam-class.}

\item{progressbar}{See BiocParalleParam-class.}

\item{RNGseed}{See BiocParalleParam-class.}

\item{manager.hostname}{character(1) host name of redis server.}

\item{manager.port}{integer(1) port of redis server.}

\item{is.worker}{logical(1) \code{bpstart()} creates worker-only
(\code{TRUE}), manager-only (\code{TRUE}), or manager and
worker (\code{NA}, default) connections.}

\item{x}{A `RedisParam` instance.}

\item{\dots}{ignored.}
}
\description{
Enable redis-based parallel evaluation in BiocParallel
}
\details{
Use an instance of `RedisParam()` for interactive parallel
    evaluation using `bplapply()` or `bpiterate()`. `RedisParam()`
    requires access to a redis server, running on
    `manager.hostname` (e.g., 127.0.0.1) on `manager.port` (e.g.,
    6379). `bplapply()` and `bpiterate()` will start and stop redis
    workers.

    It may be convenient to use `bpstart()` and `bpstop()`
    independently, to amortize the cost of worker start-up across
    multiple calls to `bplapply()` / `bpiterate()`.
}
\examples{
if (requireNamespace("redux")) {
    res <- bplapply(1:20, function(i) Sys.getpid(), BPPARAM = RedisParam())
    table(unlist(res))
}

}
