% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedisBackend-class.R
\name{RedisBackend}
\alias{RedisBackend}
\alias{.recv,RedisBackend-method}
\alias{.send,RedisBackend-method}
\alias{.close,RedisBackend-method}
\alias{.recv_any,RedisBackend-method}
\alias{.recv_all,RedisBackend-method}
\alias{.send_to,RedisBackend-method}
\alias{bpjobname,RedisBackend-method}
\alias{bpworkers,RedisBackend-method}
\title{Creating the Redis backend}
\usage{
RedisBackend(
  RedisParam = NULL,
  jobname = "",
  host = rphost(),
  port = rpport(),
  password = rppassword(),
  timeout = 2592000L,
  type = c("manager", "worker"),
  id = NULL,
  workerOffset = NULL,
  RNGseed = FALSE,
  log = FALSE
)

\S4method{.recv}{RedisBackend}(worker)

\S4method{.send}{RedisBackend}(worker, value)

\S4method{.close}{RedisBackend}(worker)

\S4method{.recv_any}{RedisBackend}(backend)

\S4method{.recv_all}{RedisBackend}(backend)

\S4method{.send_to}{RedisBackend}(backend, node, value)

\S4method{bpjobname}{RedisBackend}(x)

\S4method{bpworkers}{RedisBackend}(x)
}
\arguments{
\item{RedisParam}{RedisParam, if this argument is not NULL, all the
other arguments will be ignored except `type`.}

\item{jobname}{character(1) The job name used by the manager and
workers to connect.}

\item{host}{character(1) The host of the Redis server.}

\item{port}{integer(1) The port of the Redis server.}

\item{password}{character(1) The password of the redis server.}

\item{timeout}{integer(1) The waiting time in `BLPOP`.}

\item{type}{character(1) The type of the backend (manager or worker?).}

\item{id}{character(1) The manager/worker ID. If not given by the user and
the environment `REDISPARAM_ID` is not defined, a random ID will be used}

\item{workerOffset}{Integer(1) The starting offset of the worker, this is used
to alleviate the worker conflict when multiple managers are connected with the
same set of workers.}

\item{RNGseed}{logical(1) Whether the worker uses seed}

\item{log}{logical(1) Whether to enable the log}
}
\description{
Creating the Redis backend
}
\keyword{internal}
