% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedisBackend-class.R
\name{RedisBackend}
\alias{RedisBackend}
\alias{.recv,RedisBackend-method}
\alias{.send,RedisBackend-method}
\alias{.close,RedisBackend-method}
\alias{.recv_any,RedisBackend-method}
\alias{.recv_all,RedisBackend-method}
\alias{.send_to,RedisBackend-method}
\alias{bpjobname,RedisBackend-method}
\alias{bpworkers,RedisBackend-method}
\title{Creating the Redis backend}
\usage{
RedisBackend(
  RedisParam = NULL,
  jobname = "",
  host = rphost(),
  port = rpport(),
  password = rppassword(),
  timeout = 2592000L,
  type = c("manager", "worker"),
  id = NULL,
  workerOffset = NULL,
  seed = FALSE,
  log = FALSE
)

\S4method{.recv}{RedisBackend}(worker)

\S4method{.send}{RedisBackend}(worker, value)

\S4method{.close}{RedisBackend}(worker)

\S4method{.recv_any}{RedisBackend}(backend)

\S4method{.recv_all}{RedisBackend}(backend)

\S4method{.send_to}{RedisBackend}(backend, node, value)

\S4method{bpjobname}{RedisBackend}(x)

\S4method{bpworkers}{RedisBackend}(x)
}
\arguments{
\item{RedisParam}{RedisParam, if this argument is not NULL, all the
other arguments will be ignored except `type`.}

\item{jobname}{character(1) The job name used by the manager and
workers to connect.}

\item{host}{character(1) The host of the Redis server.}

\item{port}{integer(1) The port of the Redis server.}

\item{password}{character(1) The password of the redis server.}

\item{timeout}{integer(1) The waiting time in `BLPOP`.}

\item{type}{character(1) The type of the backend (manager or worker?).}
}
\description{
Creating the Redis backend
}
\keyword{internal}
